Refactor out ROI calculation functions into separate file
