+++ src/components/roiCalculation/index.js
@@ -1,5 +1,19 @@
 import ROICalculationPanel from "./ROICalculationPanel/ROICalculationPanel";
 import ROICalculationResults from "./ROICalculationResults/ROICalculationResults";
 import ROICalculationSurvey from "./ROICalculationSurvey/ROICalculationSurvey";
+import {
+  calculateSurveyScore,
+  calculatePercentageReturn,
+  calculateNetReturn,
+  getScoreBucketName,
+} from "./roiCalculator/roiCalculator";
 
-export { ROICalculationPanel, ROICalculationResults, ROICalculationSurvey };
+export {
+  ROICalculationPanel,
+  ROICalculationResults,
+  ROICalculationSurvey,
+  calculateSurveyScore,
+  calculatePercentageReturn,
+  calculateNetReturn,
+  getScoreBucketName,
+};
